Fix the issue with <pre> tags not getting the language- class
PrismJS is supposed to add the `language-*` class automatically after
the content is rendered but when a route is changed and a new view
is shown, Prism doesn't recognize that and just doesn't run itself.

This makes the <pre> tag not have the language class breaking the whole
code highlight part.